+++ relayer/src/transfer.rs
@@ -9,6 +9,9 @@ use ibc::tx_msg::Msg;
 use crate::chain::{Chain, CosmosSdkChain};
 use crate::config::ChainConfig;
 use crate::error::Error;
+use ibc::timestamp::Timestamp;
+use ibc::Height;
+use std::time::Duration;
 
 #[derive(Debug, Error)]
 pub enum PacketError {
@@ -33,7 +36,8 @@ pub struct TransferOptions {
     pub amount: u64,
     pub denom: String,
     pub receiver: Option<String>,
-    pub height_offset: u64,
+    pub timeout_height_offset: u64,
+    pub timeout_seconds: Duration,
     pub number_msgs: usize,
 }
 
@@ -52,9 +56,28 @@ pub fn build_and_send_transfer_messages(
         .get_signer()
         .map_err(PacketError::KeyError)?;
 
-    let latest_height = packet_dst_chain
-        .query_latest_height()
-        .map_err(|_| PacketError::Failed("Height error".to_string()))?;
+    let timeout_timestamp = if opts.timeout_seconds == Duration::from_secs(0) {
+        Timestamp::none()
+    } else {
+        Timestamp::from_nanoseconds(
+            Timestamp::now().as_nanoseconds() + opts.timeout_seconds.as_nanos() as u64,
+        )
+        .map_err(|_| {
+            PacketError::Failed(format!(
+                "invalid timeout timestamp {:?}",
+                opts.timeout_seconds
+            ))
+        })?
+    };
+
+    let timeout_height = if opts.timeout_height_offset == 0 {
+        Height::zero()
+    } else {
+        packet_dst_chain
+            .query_latest_height()
+            .map_err(|_| PacketError::Failed("Height error".to_string()))?
+            .add(opts.timeout_height_offset)
+    };
 
     let msg = MsgTransfer {
         source_port: opts.packet_src_port_id.clone(),
@@ -65,8 +88,8 @@ pub fn build_and_send_transfer_messages(
         }),
         sender,
         receiver,
-        timeout_height: latest_height.add(opts.height_offset),
-        timeout_timestamp: 0,
+        timeout_height,
+        timeout_timestamp,
     };
 
     let raw_msg = msg.to_any();